reformatted contest as per the discussion in the stand-up
Candidates table is no longer on the page, the ballot table now has sections for both the counts of referendems and candidates that ballot contains.

Fixed unit tests to support these changes, also removed e2e tests for the candidate and candidates pages till the REST services are fully implemented.