[TH] basic csv parse error surfacing
It’s not a great solution, I was hindered by the
current implementation with the child processing
and the csv parsing libraries own limitations, and
my lack of real familiarity with node.

A better, long term solution would recognize
ParseErrors as real things and collect them all up
for all files and present those like any other
data error. But that’s not possible with the
current parsing library (it only errors out once,
not for each line, and to prevent it from trying
to consolidate and fold up the broken data I had
to just stop processing all together so only get
one error for one file at a time).

Still, this is way better than what’s there.

We can investigate a similar solution for XML 
files as well.