[TH] brings error parity to XML files
The xml stream parsing library, like the csv 
one, has undocumented on(‘error’,. . .) callback
functionality, but whatever code they have in place
to do the parsed event handlers doesn’t call into
the handler properly. I would get it to fire, but
no error information was passed through. Instead,
I’m relying on a .once(‘error’,. . .) callback 
instead, and so I have to short-circuit the rest
of processing, just like with the CSV.

It’s certainly not ideal, but it should do the job
to get more information back to the states on the
inherent problems with feeds.