[TH] Swap fast-csv for node-csv
This is an attempt to handle poorly formed CSV
data files better. I picked this particular
library mostly because it seemed to be much more
popular than the fast-csv library. After
incorporating it, I was happy to find some line
numbers showing up in output indicating error spots,
however the excitement was perhaps premature when
I couldn’t detect any errors at the given line
number. It was then I noticed an undocumented
option to turn on a “relax” parsing mode, which
made it not so picky on things like unescaped quotes.
Turning this mode on helped get it to process an
entire zip file. But then I encountered other Metis
related issues after that. So I’d suggest doing
some tests with this branch and some of the
feeds we’ve gotten and see how it compares.