[CS WR] Redirect non-https clients to https
AWS load balancers add an X-Forwarded-Proto header with the protocol
used by clients. We want to redirect anyone coming over http to https.

A new /ping route has been added so that the load balancer has a page to
health check on that clients won't care about. (The health check
requests may have the X-Forwarded-Proto header set to "http" so it needs
to be able to be accessed without an https redirect.)