Fix sub, set default to nil, handle nil values
One of the things that took me longest to figure 
out was why I wasn't getting a date value after
subscribing to it. The method signature for the
sub function to get the value from the db had
the extra `_` before `db`, and surprise surprise,
there is a 2 arity version of these functions.
So rather than getting a db, we were getting
something else and attempting to use it like a
hash, which silently returned nil.

This code primarily fixes that, and then sets
the initial db value to nil (so we can force 
someone to actually set it). As such, I use the
`some->` to protect from nil values.

One more thing. When I tried returning a js/Date
from the reg-sub, it didn't like that. It seems
somewhere in the bowels of either re-frame or 
reagent it wants non-Objects. So I use `.getTime`
to return it as a numerical value, and convert
back to js/Date later.