Pass pikaday an ratom, and use the pikaday on-select
Pikaday really wants a ratom (I also tried reactions
but with little success). The thing returned by
the subscribe doesn't work as is, but wrapping it
up as a ratom when we hand it over to pikaday does
work.

Next, I hooked into the pikaday-attrs, and there
I was able to set the `:on-select` handler that it
normally defaults to push data back into the ratom.
Now it does the normal dispatch, which updates the
db, which then triggers a re-render because the
subscribed to value changes.