Adds a post endpoint to generate files to an election
URL: /earlyvote/elections/:electionid/generate

The work here is 95% done, it all works as is, but
we discovered we are missing the TZ offset on 
schedules so once we add that, we can finish out
the file generation here.

All set up for deployment to staging/production,
consul keys and S3 buckets have been setup already.