give super-admins the ability to filter files
Allows super-admin users to pick a state to
filter the files by that state, taking advantage
of the prefixing we've done.

Since we've hit a 1000 file limit, this was the
quickest way for now to both introduce a useful
feature and to give us some breathing room while
we consider a more stable long term solution.