Utilize the stats information BATT is now returning
Improved batch-address-test-tool recently to pass
back counts of each stat that was present. So using
those same stat names here to get the value if it's
there, or 0 if it's not (ie no addresses resulted
in that status). It's not 100% ideal to have to
keep the stats in sync on both applications, but
we haven't changed them ever so it's probably ok.

Now if something goes wrong with BATT and it starts
producing all "No Result"s again, it should be much
more obvious.