Switch over to VIT v2
Note, this cannot be deployed until the paired work
in GTTP/VIT to be able to configure the Civic Info
API Key in the init options is deployed, otherwise
we won't be able to override the API Key baked in
with the build with one that can access staged
data. Well, it could be deployed, it just won't
have access to staged data.